// src/app/models/entidades-financieras.ts
export class EntidadesFinancieras {
    entidadFinanciera_id: number = 0;
    nombre: string = '';
    valorCotizacionMax: number = 0;
    cuotaInicialMin: number = 0;
    TEAmin:  number = 0;
    TEAmax:  number = 0;
    precioMin:  number = 0;
    precioMax:  number = 0;
    plazoMax:  number = 0;
    plazoMin:  number = 0;
    seguroDesgravamen:  number = 0;
    seguroInmueble:  number = 0;
}